Split Filters module into directory with separate files per filter
