+++ src/Oceanostics.jl
@@ -159,7 +159,7 @@ include("KineticEnergyEquation.jl")
 include("TurbulentKineticEnergyEquation.jl")
 include("PotentialEnergyEquation.jl")
 include("FlowDiagnostics.jl")
-include("Filters.jl")
+include("Filters/Filters.jl")
 include("ProgressMessengers/ProgressMessengers.jl")
 
 using .TracerEquation, .TracerVarianceEquation, .KineticEnergyEquation, .TurbulentKineticEnergyEquation, .PotentialEnergyEquation
